Rename internal rust_type module to avoid clash with rust_type macro
